refactor(coderd/x/nats): merge options.go into pubsub.go
